services/f-api/requirements.txt
fastapi>=0.109.2
pydantic>=2.6.1
uvicorn>=0.27.1
joblib>=1.3.2
starlette>=0.36.3
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
finnhub-python>=2.4.19
pandas>=2.2.1
motor>=3.3.2
asyncio>=3.4.3
pymongo>=4.6.2
httpx>=0.27.0
transformers>=4.38.2
